Bugs fixing

+++ Client/Client.cpp
@@ -153,15 +153,15 @@ void Client::quitAllChannels() {
     }
 
     size_t channel_count = joined_channels.size();
-    for (int i = channel_count - 1; i >= 0; --i) {
+    for (size_t i = channel_count - 1; i >= 0; --i) {
         server->log("[CLIENT] Client " + std::to_string(fd) + " quitting channel " + joined_channels[i]);
 
-            Channel* channel = server->getChannel(joined_channels[i]);
-            if (channel && channel->isMember(*this)) {
-                std::string quit_message = "PART " + channel->getName();
-                channel->broadcastToMembers(*this, quit_message);
-                channel->removeMember(*this);
-            }
+        Channel* channel = server->getChannel(joined_channels[i]);
+        if (channel && channel->isMember(*this)) {
+            std::string quit_message = "PART " + channel->getName();
+            channel->broadcastToMembers(*this, quit_message);
+            channel->removeMember(*this);
+        }
     }
 
     joined_channels.clear();
